Concentrate import edit mechanics behind one private Module
Lexer initialization, URL masking, range-based replacement, and import attribute range handling now share one implementation while existing entrypoints remain shims.

Constraint: Attribute and HTTP URL position behavior must remain byte-compatible

Rejected: Regex-only import attributes | string literals and generated bundles require lexer-bounded edits

Confidence: high

Scope-risk: moderate

Tested: RED deno test --no-check --allow-all src/transforms/import-rewriter/core.test.ts failed with module not found for import-edit.ts

Tested: GREEN deno test --no-check --allow-all src/transforms/import-rewriter/core.test.ts src/transforms/esm/import-attributes.test.ts

Tested: deno check src/transforms/import-rewriter/import-edit.ts src/transforms/import-rewriter/parse-cache.ts src/transforms/esm/import-attributes.ts src/transforms/import-rewriter/core.test.ts

Tested: deno fmt --check src/transforms/import-rewriter/import-edit.ts src/transforms/import-rewriter/parse-cache.ts src/transforms/esm/import-attributes.ts src/transforms/import-rewriter/core.test.ts

Tested: git diff --check
